+++ source/draw.c
@@ -143,14 +143,14 @@ HB_FUNC(HWG_PIE)
 {
   int32_t res = Pie(hwg_par_HDC(1), hwg_par_int(2), hwg_par_int(3), hwg_par_int(4), hwg_par_int(5), hwg_par_int(6),
                 hwg_par_int(7), hwg_par_int(8), hwg_par_int(9));
-  hb_retnl(res ? 0 : (LONG)GetLastError()); // TODO: o retorno da função é BOOL
+  hb_retnl(res ? 0 : (int32_t)GetLastError()); // TODO: o retorno da função é BOOL
 }
 
 // HWG_ELLIPSE(HDC, nLeft, nTop, nRight, nBottom) --> numeric
 HB_FUNC(HWG_ELLIPSE)
 {
   int32_t res = Ellipse(hwg_par_HDC(1), hwg_par_int(2), hwg_par_int(3), hwg_par_int(4), hwg_par_int(5));
-  hb_retnl(res ? 0 : (LONG)GetLastError()); // TODO: o retorno da função é BOOL
+  hb_retnl(res ? 0 : (int32_t)GetLastError()); // TODO: o retorno da função é BOOL
 }
 
 // HWG_FILLRECT(HDC, nLeft, nTop, nRight, nBottom, HBRUSH) -->
@@ -321,7 +321,7 @@ HB_FUNC(HWG_WINDOW2BITMAP)
 
   DeleteDC(hDCmem);
   DeleteDC(hDC);
-  // hb_retnl((LONG)hBitmap);
+  // hb_retnl((int32_t)hBitmap);
   hwg_ret_HBITMAP(hBitmap);
 }
 
@@ -714,7 +714,7 @@ HB_FUNC(HWG_GETDRAWITEMINFO)
   hb_itemArrayPut(aMetr, 8, temp);
   hb_itemRelease(temp);
 
-  temp = hb_itemPutNL(HWG_NULLPTR, (LONG)lpdis->itemState);
+  temp = hb_itemPutNL(HWG_NULLPTR, (int32_t)lpdis->itemState);
   hb_itemArrayPut(aMetr, 9, temp);
   hb_itemRelease(temp);
